requirements.txt
langchain==0.2.2
langchain_community==0.2.2
langchain_core==0.2.4
langchain_google_genai==1.0.5
langchain_text_splitters==0.2.1
tiktoken==0.7.0
pypdf==4.2.0
python-dotenv==1.0.1
streamlit_option_menu_patch==0.3.12
supabase==2.5.1
